# Optimize DateTime.ToString() to use a cached value on multiple calls.