**Before:**
Completed 261,312 operations in a weighted-average of 10s (26,103.550858 ops/s, 3.8309e-05 s/op)

**After:**
Completed 10,024,257 operations in a weighted-average of 10s (1,001,638.652412 ops/s, 9.98364e-07 s/op)

Caching results in **38x improvement**