@ahsonkhan 2024-08-20
Exploration/discussion complete, including https://github.com/Azure/azure-sdk-for-cpp/pull/5925

For `DateTime`, specifically, there are some downsides to caching:
- Removing const on DateTime’s ToString() would be considered breaking, we'd have to add the “optimized” non-const overloads.
- The `ToString()` methods accept different format enums, resulting in multiple potential string values, all of which would need to be cached as independent versions. That bloats the size of the DateTime struct too much.
- Currently `DateTime` can be used as a constexpr. The moment it has std::string members, it can’t be (`std::string` doesn’t support it until C++20). We’d end up having to use a constant sized char array for the caching.
- To lazy evaluate the `ToString()` cache, we'd need to add a mutex, to continue supporting immutability in a multi-threaded scenario. Alternatively, we'd have to build the string in the `DateTime` ctor. (Thanks @LarryOsterman).
- We'd want performance evaluation to understand if SDK/customer intended usage results in repeated calls to `ToString()` on the same DateTime, and whether that's a bottleneck. That would drive the need for optimizing the type, over having a simpler implementation.